+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project are documented in this file.
 - **retry layer**: Integrate retry logic with task handling ([#512](https://github.com/geofmureithi/apalis/pull/512))
 - **generic retry**: Persist check for tasks ([#498](https://github.com/geofmureithi/apalis/pull/498))
 - **native TLS**: Add `async-std-comp-native-tls` and `tokio-comp-native-tls` features ([#525](https://github.com/geofmureithi/apalis/pull/525))
+- **cron** : Introduce CronContext ([#488](https://github.com/geofmureithi/apalis/pull/488))
 
 ### Fixed
 - **PostgresStorage**: PostgresStorage get_jobs status conditional ([#524](https://github.com/geofmureithi/apalis/pull/524))